Pluralize/rename typo'd filename.
done as part of CURA-10994
