+++ tests/Settings/TestAppendAdditionalSettings.py
@@ -5,7 +5,7 @@ import os
 import pytest
 from unittest.mock import MagicMock, patch
 
-from UM.Settings.AdditionalSettingDefinitionAppender import AdditionalSettingDefinitionsAppender
+from UM.Settings.AdditionalSettingDefinitionsAppender import AdditionalSettingDefinitionsAppender
 from UM.Settings.DefinitionContainer import DefinitionContainer
 from UM.VersionUpgradeManager import VersionUpgradeManager
